Refactor signal websocket
