+++ server/src/managers/websocket/mod.rs
@@ -1,2 +1,3 @@
 pub mod connection;
 pub mod websocket_manager;
+pub mod signal_websocket;
